Move generate share file token logic into share controller
